Linked, recursion, and trees notes added
